In game menu template
Once a piece of evidence or a transcript is presented, it will appear in the corner of the screen for all the users currently within the courtroom. From there, if they wish to view the image or transcript full screen, they can press a button on their controller to toggle the evidence or transcript full screen. If a piece of evidence and transcript are being presented at the same time, they will appear in opposite bottom corners and the buttons on the side of the controller can be used to select which one to view individually.