Texturing + UVMapping Changes
Updated Textures on the Pew-Backs, and the entirety of the judges area. Some textures there might need to be edited slightly, but that can be done at a later date, simply a drag and drop since the UVMapping is finished on those too. Lady Liberty and her sword and weights have also been fully textured and mapped. Arches, inner walls and outer walls have also been finished. However slight colour changes can be made in the future also.