Textures
Changed the texture file type for the prince of whales plaque (.tifs dont work in unreal). Successfully started up adding overlay textures to the walls in the project. However only standard textures are able to be imported into unreal. A different method is needed. Updated the .fbx file showing changes.